medi_care/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-slugify==8.0.1
Werkzeug==3.0.0
gunicorn==21.2.0
# optional for selenium
selenium==5.9.0
webdriver-manager==4.11.0
